Removed old tests.
