+++ opensourceleg/time/time.py
@@ -119,10 +119,7 @@ class SoftRealtimeLoop:
             print("\tavg error: %.3f milliseconds" % (1e3 * self.sum_err / self.n))
             print(
                 "\tstddev error: %.3f milliseconds"
-                % (
-                    1e3
-                    * sqrt((self.sum_var - self.sum_err**2 / self.n) / (self.n - 1))
-                )
+                % (1e3 * sqrt((self.sum_var - self.sum_err**2 / self.n) / (self.n - 1)))
             )
             print(
                 "\tpercent of time sleeping: %.1f %%"
