--- tests/test_logger/test_log_stream_level
@@ -1,1 +1,0 @@
-[2023-06-12 11:58:40,376] WARNING: Invalid logging level: BAD LEVEL
