+++ tests/test_math/test_math.py
@@ -1,10 +1,49 @@
 import pytest
 
-from opensourceleg.math import math
+from opensourceleg.math import EdgeDetector, SaturatingRamp, ThermalModel
+
+
+def test_edge_detector_init():
+
+    edi = EdgeDetector(bool_in=False)
+    assert edi.cur_state == False
+    assert edi.rising_edge == False
+    assert edi.falling_edge == False
+
+
+def test_edge_detector_update():
+
+    edu = EdgeDetector(bool_in=False)
+    edu.update(bool_in=True)
+    assert edu.rising_edge == True
+    assert edu.falling_edge == False
+    assert edu.cur_state == True
+    edu2 = EdgeDetector(bool_in=True)
+    edu2.update(bool_in=False)
+    assert edu2.rising_edge == False
+    assert edu2.falling_edge == True
+    assert edu2.cur_state == False
+
+
+def test_saturating_ramp_init():
+
+    sri = SaturatingRamp(loop_frequency=100, ramp_time=1.0)
+    assert sri.delta_per_update == 1.0 / 100
+    assert sri.value == 0.0
+
+
+def test_saturating_ramp_update():
+
+    sru = SaturatingRamp(loop_frequency=100, ramp_time=1.0)
+    sru.update(enable_ramp=True)
+    assert sru.value == 0.01
+    sru.update(enable_ramp=False)
+    assert sru.value == 0.0
+
 
 # Initializing the ThermalModel objects
-test_model_default = math.ThermalModel()
-test_model_specified = math.ThermalModel(
+test_model_default = ThermalModel()
+test_model_specified = ThermalModel(
     ambient=10,
     temp_limit_windings=100,
     soft_border_C_windings=25,
